fix(lib): add lenient fallbacks for _all variant
This commit ensures that when calling connected_displays_all, if a call
to QueryDisplayConfig fails, we generate an empty device_info_map.

This map is only used to add output technology metadata to the resuting
device output, so Device.output_technology has been made an Option<T>
and will be populated with None in the rare occasions on very specific
VDI configurations that this call fails.

Similarly, if get_display_devices_from_hmonitor returns an empty Vec, we
retry with get_display_devices_from_hmonitor_lenient which uses the old
komorebi-style monitor info lookup strategy confirmed to work with VDIs
that are now seeing problems.