fix(lib): continue exec for all displays on query_display_config failures
This commit ensures that when calling connected_displays_all, if a call
to QueryDisplayConfig fails, we generate an empty device_info_map.

This map is only used to add output technology metadata to the resuting
device output, so Device.output_technology has been made an Option<T>
and will be populated with None in the rare occasions on very specific
VDI configurations that this call fails.