fix: revert ignore empty path and set serial to None
This commit reverts the previous commit that ignored devices with empty
path. Apparently some devices might have an empty path and still need to
be handled, like remote displays. Now to fix the panics when trying to
get the `serial_number_id` on devices with empty paths we handle that on
the `try_with_serial_number_id` function it self by immediately setting
`serial_number_id` to `None` and exiting early when the `device_path` is
empty.